Extract containment narrowing helpers
